requirements.txt
autopep8==1.5.5
backcall==0.2.0
bcrypt==3.2.0
cffi==1.14.5
decorator==4.4.2
googleapis-common-protos==1.52.0
grpc-interceptor==0.13.0
grpcio==1.36.0
grpcio-status==1.35.0
grpcio-testing==1.36.0
grpcio-tools==1.35.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
protobuf==3.15.0
ptyprocess==0.7.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.8.0
PyJWT==2.0.1
pymongo==3.11.3
python-dateutil==2.8.1
pyzmq==22.0.3
six==1.15.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
